--- library/src/lib/calendar/i18n/calendar-i18n-default.ts
@@ -1,36 +1,0 @@
-import { Inject, Injectable, LOCALE_ID } from '@angular/core';
-import { CalendarI18n } from './calendar-i18n';
-import { FormStyle, getLocaleDayNames, getLocaleMonthNames, TranslationWidth } from '@angular/common';
-
-@Injectable()
-export class CalendarI18nDefault extends CalendarI18n {
-
-    private weekdaysShort: string[];
-    private monthsShort: string[];
-    private monthsFull: string[];
-
-    constructor(@Inject(LOCALE_ID) private locale: string) {
-        super();
-        const sundayStartWeekdays = getLocaleDayNames(locale, FormStyle.Standalone, TranslationWidth.Short);
-        this.weekdaysShort = sundayStartWeekdays.map((day, index) => sundayStartWeekdays[(index + 1) % 7]);
-        this.monthsShort = getLocaleMonthNames(locale, FormStyle.Standalone, TranslationWidth.Abbreviated);
-        this.monthsFull = getLocaleMonthNames(locale, FormStyle.Standalone, TranslationWidth.Wide);
-    }
-
-    getDayAriaLabel(date: Date): string {
-        return date.toLocaleDateString();
-    }
-
-    getMonthAbbreviatedName(month: number, year?: number): string {
-        return this.monthsShort[month - 1];
-    }
-
-    getMonthFullName(month: number, year?: number): string {
-        return this.monthsFull[month - 1];
-    }
-
-    getWeekdayAbbreviatedName(weekday: number): string {
-        return this.weekdaysShort[weekday - 1];
-    }
-
-}
