+++ tslint.yaml
@@ -55,7 +55,7 @@ rules:
     no-switch-case-fall-through: true
     no-unnecessary-initializer: true
     no-unused-expression: true
-    no-use-before-declare: true
+    no-use-before-declare: false
     no-var-keyword: true
     object-literal-sort-keys: false
     one-line:
